enauto1/m3/requirements.txt
netmiko
jinja2
pyyaml
